+++ src/main/java/com/project/storemanager_api/service/PaymentService.java
@@ -15,17 +15,12 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDate;
-import java.time.YearMonth;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 import static com.project.storemanager_api.domain.order.entity.Order.OrderStatus.SUCCESS;
-import static com.project.storemanager_api.domain.pay.entity.PaymentDetail.PaymentType;
-import static com.project.storemanager_api.domain.pay.entity.PaymentDetail.PaymentType.CASH;
 
 @Service
 @Slf4j
@@ -69,8 +64,6 @@ public class PaymentService {
         // 주문 상세정보 저장
         for (CreatePaymentDetailDto payDetail : dto.getPayList()) {
             // payDetail.getExpiryDate() 날짜 확인
-            checkExpiryDate(payDetail.getExpiryDate(), payDetail.getPaymentType());
-
             paymentDetailService.savePayInfo(payDetail, generatedPaymentId);
         }
         // 결제에 사용된 카드정보 저장
@@ -160,46 +153,6 @@ public class PaymentService {
         }
 
         return menuList;
-    }
-
-    private void checkExpiryDate(String expiryDate, PaymentType paymentType) {
-        // 현금 결제는 패스
-        if (paymentType.equals(CASH)) {
-            return;
-        }
-        // 문자 형식 검증
-        validateExpiryDate(expiryDate);
-
-        // 예시 -> "2025/03"은 2025년 3월 1일로 간주
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy/MM");
-        YearMonth yearMonth = YearMonth.parse(expiryDate, formatter);
-
-        // ✅ 해당 월의 1일을 LocalDate로 변환
-        LocalDate expirationDate = yearMonth.atDay(1);
-
-        // 현재 날짜 가져오기
-        LocalDate today = LocalDate.now();
-
-        // 만료 여부 확인
-        if (today.isAfter(expirationDate)) {
-           throw new PaymentException(ErrorCode.DATE_EXPIRATION, ErrorCode.DATE_EXPIRATION.getMessage() + " " + expiryDate);
-        }
-    }
-
-    // 문자 형식 검증
-    private void validateExpiryDate(String expiryDate) {
-        if (expiryDate.length() != 7) {
-            throw new PaymentException(ErrorCode.NOT_CORRECT_EXPIRATION, ErrorCode.NOT_CORRECT_EXPIRATION.getMessage());
-        }
-        String[] parts = expiryDate.split("/"); // ["2026", "03"]
-        if (parts.length != 2) {
-            throw new PaymentException(ErrorCode.NOT_CORRECT_EXPIRATION, ErrorCode.NOT_CORRECT_EXPIRATION.getMessage());
-        }
-
-        int month = Integer.parseInt(parts[1]); // "03" → 3
-        if (month < 1 || month > 12) {
-            throw new PaymentException(ErrorCode.NOT_CORRECT_MONTH, ErrorCode.NOT_CORRECT_MONTH.getMessage());
-        }
     }
 
 }
